made form into a component
